big refactor

+++ image.h
@@ -1,4 +1,4 @@
 #include "stiv.h"
 
 void get_img_size(Image *img);
-void reduce_img_size(Image *img, uint new_w);
+void reduce_img_size(Image *img, double new_w);
